requirements.txt
beautifulsoup4==4.11.1
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.1.3
cssselect==1.1.0
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
html5lib==1.1
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
lxml==4.8.0
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tinysegmenter==0.3
tldextract==3.2.1
tqdm==4.64.0
urllib3==1.26.9
webencodings==0.5.1
xlrd==2.0.1
xmltodict==0.12.0
